LayoutTests/platform/mac-wk2/plugins/mouse-events-scaled-iframe-expected.txt

CONSOLE MESSAGE: PLUGIN: mouseDown at (300, 300)
CONSOLE MESSAGE: PLUGIN: mouseUp at (300, 300)
CONSOLE MESSAGE: PLUGIN: mouseDown at (200, 200)
CONSOLE MESSAGE: PLUGIN: mouseUp at (200, 200)
CONSOLE MESSAGE: PLUGIN: mouseDown at (133, 133)
CONSOLE MESSAGE: PLUGIN: mouseUp at (133, 133)
